+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Changed
 - Reduced logging for PGP Timestamping Server mail handling
+- Updated gnupg config documentation
 
 
 # 0.9.4 - 2020-05-09